Extracts Stub callbacks into class
